app/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
gunicorn
psycopg2-binary
boto3
django-storages
